Move container and list view related files.
