+++ src/analysis/ContextBar.ts
@@ -58,7 +58,7 @@ import {
   isAzureWorkspace,
   isGoogleWorkspace,
 } from 'src/libs/workspace-utils';
-import { StorageDetails } from 'src/pages/workspaces/hooks/useWorkspace';
+import { StorageDetails } from 'src/workspaces/hooks/useWorkspace';
 
 const contextBarStyles: { [label: string]: CSSProperties } = {
   contextBarContainer: {
