+++ src/pages/workspaces/workspace/Dashboard/WorkspaceDashboard.test.tsx
@@ -1,9 +1,9 @@
 import { screen } from '@testing-library/react';
 import React from 'react';
-import { InitializedWorkspaceWrapper as Workspace, StorageDetails } from 'src/pages/workspaces/hooks/useWorkspace';
 import { WorkspaceDashboard } from 'src/pages/workspaces/workspace/Dashboard/WorkspaceDashboard';
 import { renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultGoogleBucketOptions, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
+import { InitializedWorkspaceWrapper as Workspace, StorageDetails } from 'src/workspaces/hooks/useWorkspace';
 
 describe('WorkspaceDashboard', () => {
   const storageDetails: StorageDetails = {
