+++ src/pages/workspaces/workspace/Dashboard/WorkspaceDashboardPage.ts
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import * as breadcrumbs from 'src/components/breadcrumbs';
 import { WorkspaceDashboard } from 'src/pages/workspaces/workspace/Dashboard/WorkspaceDashboard';
-import { wrapWorkspace } from 'src/pages/workspaces/workspace/WorkspaceContainer';
+import { wrapWorkspace } from 'src/workspaces/container/WorkspaceContainer';
 
 export interface WorkspaceDashboardPageProps {
   namespace: string;
