+++ src/workspaces/container/WorkspaceAttributeNotice.test.ts
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { axe } from 'jest-axe';
 import { h } from 'react-hyperscript-helpers';
 import { WorkspaceAccessLevel } from 'src/libs/workspace-utils';
-import { WorkspaceAttributeNotice } from 'src/pages/workspaces/workspace/WorkspaceAttributeNotice';
+import { WorkspaceAttributeNotice } from 'src/workspaces/container/WorkspaceAttributeNotice';
 
 describe('WorkspaceAttributeNotice', () => {
   it('renders read-only, locked, and protected workspace attributes, with no accessibility issues', async () => {
