+++ src/workspaces/container/WorkspaceAttributeNotice.ts
