+++ src/workspaces/container/WorkspaceContainer.test.ts
@@ -5,10 +5,10 @@ import { Ajax } from 'src/libs/ajax';
 import { goToPath } from 'src/libs/nav';
 import { workspacesStore, workspaceStore } from 'src/libs/state';
 import { WorkspaceWrapper } from 'src/libs/workspace-utils';
-import { InitializedWorkspaceWrapper } from 'src/pages/workspaces/hooks/useWorkspace';
-import { WorkspaceContainer } from 'src/pages/workspaces/workspace/WorkspaceContainer';
 import { asMockedFn, renderWithAppContexts as render } from 'src/testing/test-utils';
 import { defaultAzureWorkspace, defaultGoogleWorkspace } from 'src/testing/workspace-fixtures';
+import { WorkspaceContainer } from 'src/workspaces/container/WorkspaceContainer';
+import { InitializedWorkspaceWrapper } from 'src/workspaces/hooks/useWorkspace';
 
 type NavExports = typeof import('src/libs/nav');
 jest.mock(
@@ -53,9 +53,9 @@ jest.mock<StateExports>(
   })
 );
 
-type WorkspaceMenuExports = typeof import('src/pages/workspaces/workspace/WorkspaceMenu');
-jest.mock<WorkspaceMenuExports>('src/pages/workspaces/workspace/WorkspaceMenu', () => ({
-  ...jest.requireActual('src/pages/workspaces/workspace/WorkspaceMenu'),
+type WorkspaceMenuExports = typeof import('src/workspaces/container/WorkspaceMenu');
+jest.mock<WorkspaceMenuExports>('src/workspaces/container/WorkspaceMenu', () => ({
+  ...jest.requireActual('src/workspaces/container/WorkspaceMenu'),
   WorkspaceMenu: jest.fn().mockReturnValue(null),
 }));
