+++ src/workspaces/container/WorkspaceDeletingBanner.ts
