+++ src/workspaces/list/WorkspacesList.ts
@@ -13,21 +13,18 @@ import { useOnMount } from 'src/libs/react-utils';
 import { elements as StyleElements } from 'src/libs/style';
 import { newTabLinkProps } from 'src/libs/utils';
 import { cloudProviderTypes, WorkspaceWrapper as Workspace } from 'src/libs/workspace-utils';
-import { useDeletionPolling } from 'src/pages/workspaces/hooks/useDeletionPolling';
-import { useWorkspacesWithSubmissionStats } from 'src/pages/workspaces/hooks/useWorkspacesWithSubmissionStats';
-import { categorizeWorkspaces } from 'src/pages/workspaces/WorkspacesList/CategorizedWorkspaces';
-import { RecentlyViewedWorkspaces } from 'src/pages/workspaces/WorkspacesList/RecentlyViewedWorkspaces';
+import { useDeletionPolling } from 'src/workspaces/hooks/useDeletionPolling';
+import { useWorkspacesWithSubmissionStats } from 'src/workspaces/hooks/useWorkspacesWithSubmissionStats';
+import { categorizeWorkspaces } from 'src/workspaces/list/CategorizedWorkspaces';
+import { RecentlyViewedWorkspaces } from 'src/workspaces/list/RecentlyViewedWorkspaces';
 import {
   getWorkspaceFiltersFromQuery,
   WorkspaceFilters,
   WorkspaceFilterValues,
-} from 'src/pages/workspaces/WorkspacesList/WorkspaceFilters';
-import { WorkspacesListModals } from 'src/pages/workspaces/WorkspacesList/WorkspacesListModals';
-import { WorkspacesListTabs } from 'src/pages/workspaces/WorkspacesList/WorkspacesListTabs';
-import {
-  WorkspaceUserActions,
-  WorkspaceUserActionsContext,
-} from 'src/pages/workspaces/WorkspacesList/WorkspaceUserActions';
+} from 'src/workspaces/list/WorkspaceFilters';
+import { WorkspacesListModals } from 'src/workspaces/list/WorkspacesListModals';
+import { WorkspacesListTabs } from 'src/workspaces/list/WorkspacesListTabs';
+import { WorkspaceUserActions, WorkspaceUserActionsContext } from 'src/workspaces/list/WorkspaceUserActions';
 
 export const persistenceId = 'workspaces/list';
